+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Added
+- add evaluation of `Expr` in `Solution`.
+
 ## 3.0.4 - 2020-10-30
 ### Added
 - add more SCIP functions: `getNTotalNodes`, `getNIntVars`, `getNBinVars`